+++ src/renderer/components/topology/TopologySidebar.tsx
@@ -29,7 +29,7 @@ import {
 } from '../../utils/ipc/workspaceIpcUtils';
 import { openDialog } from '../../utils/ipc/dialogIpcUtils';
 import { TOP_NAVBAR_HEIGHT } from '../MainNavbar';
-import TopologyLibrary from './TopologyLibrary';
+import TopologyLibrary from './Library/TopologyLibrary';
 import {
   setFileObjects,
   setMapObjectTypeCollection,
